added a whole bunch f test for all the binary trees, first time using pytest, still a bit shaky

+++ src/pytree/binarytree/_tree.py
@@ -1,7 +1,8 @@
 from typing import Generic, Union, Tuple, List
 from types import MethodType
+from typing_extensions import Self
 
-from ._type_hint import CT, N, Node
+from ._type_hint import CT, BN
 
 
 class BinaryTree(Generic[CT]):
@@ -13,7 +14,7 @@ class BinaryTree(Generic[CT]):
     all binary search tree variations should inherit this class
     to obtain all the necessary interface functions
     '''
-    _node_type: N = None
+    _node_type: BN = None
 
     def __init__(self):
         if self._node_type is None:
@@ -31,15 +32,7 @@ class BinaryTree(Generic[CT]):
 
     @property
     def height(self) -> int:
-        '''recursively get the height of the tree '''
-
-        def traversal_counter(node) -> int:
-            if node is None:
-                return -1
-            return 1 + max(traversal_counter(node.left),
-                           traversal_counter(node.right))
-
-        return traversal_counter(self.root)
+        return self.root.height
 
     @property
     def is_complete(self) -> bool:
@@ -87,92 +80,15 @@ class BinaryTree(Generic[CT]):
             # check whether the left & right node is perfect
             # and whether the height the node is perfect
             perfectness_check = (left_check[0] and right_check[0]
-                                 and abs(left_check[1] - right_check[1]) == 0)
+                                 and abs(left_check[1] - right_check[1]) == 0)  # noqa
 
             return (perfectness_check, 1 + max(left_check[1], right_check[1]))
 
         return traversal_check(self.root)[0]
 
     @property
-    def is_binary(self) -> bool:
-        '''
-        check whether the tree obeys the binary search tree's invariant
-        i.e:
-        - left node's value < node's value
-        - right node's value > node's value
-        '''
-
-        def traversal_check(node) -> bool:
-            # keep going down the chain of nodes
-            # until the leftmost/rightmost node has been reached
-            # then, return True, as leaf nodes has no child nodes
-            if node is None:
-                return True
-
-            left_check = traversal_check(node.left)
-            right_check = traversal_check(node.right)
-
-            # check whether the left & right node obey the BST invariant
-            check_binary = left_check and right_check
-
-            # then, check the node itself, whether it obeys the BST invariant
-            if node.left and node.left >= node:
-                check_binary = False
-            if node.right and node.right < node:
-                check_binary = False
-
-            return check_binary
-
-        return traversal_check(self.root)
-
-    @property
-    def is_balanced(self) -> bool:
-        '''
-        check whether the tree is balanced, i.e both side of the tree,
-        left & right have similar/same number of nodes
-        -> the difference in number of nodes for both side
-           of every node does not exceed 1
-
-        STEPS:
-
-            1. go down until the leftmost & rightmost node has been reached
-            2. start checking whether each node is balanced
-            3. return their height along with whether they're balanced or not
-            4. unfold the recursion to the prev node
-            5. rinse & repeat until the recursion unfolds
-               back to the starting node/root node
-
-            * so, if any one of the node, starting from the leaf nodes,
-            is unbalanced it will cause will the other nodes 'above'
-            him to be unbalanced as well due to all to them depending on
-            the last node's balance_value(boolean)
-
-            * basically, only 2 value is passed around,
-              the balance_value & the height of the node
-              - the balance_value is required for the said chain reaction
-              - while the node's height is required for the checking
-        '''
-
-        def traversal_check(node) -> Tuple[bool, int]:
-            # keep going down the chain of nodes
-            # until the leftmost/rightmost node has been reached
-            # return True, as leaf nodes has no child nodes and are balanced
-            # + the height of the leaf node, which is -1 since it has no child
-            if node is None:
-                return (True, -1)
-
-            left_height = traversal_check(node.left)
-            right_height = traversal_check(node.right)
-
-            # check whether the left & right node is balanced
-            # and whether the height the node is balanced
-            balanced = (left_height[0] and right_height[0]
-                        and abs(left_height[1] - right_height[1]) <= 1)
-
-            # return the 'balanced' variable and the height of the current node
-            return (balanced, 1 + max(left_height[1], right_height[1]))
-
-        return traversal_check(self.root)[0]
+    def is_empty(self) -> bool:
+        return self.root.value is None
 
     @classmethod
     def fill_tree(cls, values: List[CT]) -> 'BinaryTree':
@@ -182,15 +98,20 @@ class BinaryTree(Generic[CT]):
             new_bst.insert(value)
         return new_bst
 
+    def extend(self, values: List[CT]) -> None:
+        '''generates a binary tree with all the values from a list'''
+        for value in values:
+            self.insert(value)
+
     def insert(self, value: CT) -> None:
         '''add a node with the given value into the tree'''
         if self.root.value is None:
             self.root.value = value
-
         else:
-            new_root = self.root.insert_node(value)
-            if new_root:
-                self.root = new_root
+            self.root.insert_node(value)
+
+        if self.root.parent is not None:
+            self.root = self.root.get_root()
 
         self._size += 1
 
@@ -199,15 +120,18 @@ class BinaryTree(Generic[CT]):
         if self.root.value is None:
             raise ValueError(f'{value} is not in {self.__class__.__name__}')
 
+        if not isinstance(value, type(self.root.value)):
+            raise TypeError(f"tree does not contain value of type {type(value).__name__}")
+
         node_to_delete = self.root.find_node(value)
 
         if node_to_delete is None:
             raise ValueError(f'{value} is not in {self.__class__.__name__}')
 
-        new_root = self.root.delete_node(node_to_delete)
+        self.root.delete_node(node_to_delete)
 
-        if new_root:
-            self.root = new_root
+        if self.root.parent is not None:
+            self.root = self.root.get_root()
 
         self._size -= 1
 
@@ -233,8 +157,9 @@ class BinaryTree(Generic[CT]):
         popping_func = popping_options['min']
 
         if value:
-            popping_func = lambda: popping_options['val'](value)  # noqa: E731
-        if key:
+            def popping_func():
+                return popping_options['val'](value)  # noqa: E731
+        elif key:
             popping_func = popping_options[key]
 
         found_val = popping_func()
@@ -247,10 +172,11 @@ class BinaryTree(Generic[CT]):
         self.root.left = None
         self.root.right = None
         self.root.value = None
+        self._size = 0
 
     def traverse(self,
                  key: str = 'in',
-                 node: bool = False) -> List[Union[Node, CT]]:
+                 node: bool = False) -> List[Union[BN, CT]]:
         '''
         returns list of all the items in the tree in the given order type
         in-order  ['in']: from min-to-max
@@ -262,8 +188,12 @@ class BinaryTree(Generic[CT]):
             return []
         return self.root.traverse_node(key, node)
 
-    def find(self, value: CT, node: bool = False) -> Union[Node, CT]:
+    def find(self, value: CT, node: bool = False) -> Union[BN, CT]:
         '''get the node with the given value'''
+        if self.root.value is None:
+            return None
+        if not isinstance(value, type(self.root.value)):
+            raise TypeError(f"tree does not contain value of type '{type(value).__name__}'")
         target_node = self.root.find_node(value)
         if target_node:
             return target_node.value if not node else target_node
@@ -271,8 +201,12 @@ class BinaryTree(Generic[CT]):
     def find_lt(self,
                 value: CT,
                 node: bool = False,
-                **kwargs) -> Union[Node, CT]:
+                **kwargs) -> Union[BN, CT]:
         '''get the node with the given value'''
+        if self.root.value is None:
+            return None
+        if not isinstance(value, type(self.root.value)):
+            raise TypeError(f"tree does not contain value of type '{type(value).__name__}'")
         target_node = self.root.find_lt_node(value, **kwargs)
         if target_node:
             return target_node.value if not node else target_node
@@ -280,8 +214,12 @@ class BinaryTree(Generic[CT]):
     def find_gt(self,
                 value: CT,
                 node: bool = False,
-                **kwargs) -> Union[Node, CT]:
+                **kwargs) -> Union[BN, CT]:
         '''find the node with the closest value that's > the given value'''
+        if self.root.value is None:
+            return None
+        if not isinstance(value, type(self.root.value)):
+            raise TypeError(f"tree does not contain value of type '{type(value).__name__}'")
         target_node = self.root.find_gt_node(value, **kwargs)
         if target_node:
             return target_node.value if not node else target_node
@@ -289,8 +227,12 @@ class BinaryTree(Generic[CT]):
     def find_le(self,
                 value: CT,
                 node: bool = False,
-                **kwargs) -> Union[Node, CT]:
+                **kwargs) -> Union[BN, CT]:
         '''get the node with the given value'''
+        if self.root.value is None:
+            return None
+        if not isinstance(value, type(self.root.value)):
+            raise TypeError(f"tree does not contain value of type '{type(value).__name__}'")
         target_node = self.root.find_le_node(value, **kwargs)
         if target_node:
             return target_node.value if not node else target_node
@@ -298,19 +240,23 @@ class BinaryTree(Generic[CT]):
     def find_ge(self,
                 value: CT,
                 node: bool = False,
-                **kwargs) -> Union[Node, CT]:
+                **kwargs) -> Union[BN, CT]:
         '''find the node with the closest value that's < the given value'''
+        if self.root.value is None:
+            return None
+        if not isinstance(value, type(self.root.value)):
+            raise TypeError(f"tree does not contain value of type '{type(value).__name__}'")
         target_node = self.root.find_ge_node(value, **kwargs)
         if target_node:
             return target_node.value if not node else target_node
 
-    def find_max(self, node: bool = False, **kwargs) -> Union[Node, CT]:
+    def find_max(self, node: bool = False, **kwargs) -> Union[BN, CT]:
         '''get the node with the maximum value in the tree'''
         target_node = self.root.find_max_node(**kwargs)
         if target_node:
             return target_node.value if not node else target_node
 
-    def find_min(self, node: bool = False, **kwargs) -> Union[Node, CT]:
+    def find_min(self, node: bool = False, **kwargs) -> Union[BN, CT]:
         '''get the node with the minimum value in the tree'''
         target_node = self.root.find_min_node(**kwargs)
         if target_node:
@@ -325,8 +271,7 @@ class BinaryTree(Generic[CT]):
                 raise TypeError(
                     f"cannot add '{type(self).__name__}({self.dtype.__name__})' \
                       with '{type(other).__name__}({other.dtype.__name__})'")
-            return type(self).fill_tree([val for val in self] +
-                                        [val for val in other])
+            return type(self).fill_tree([val for val in self] + [val for val in other])
 
         try:
             return type(self).fill_tree(self.traverse()).insert(other)
@@ -401,30 +346,37 @@ class BinaryTree(Generic[CT]):
                 f"cannot delete value of type '{other.__class__.__name__}' \
                 from '{self.__class__.__name__}({self.dtype.__name__})'")
 
-    def __getattribute__(self, attr_name):
-        attr = super().__getattribute__(attr_name)
+    def is_subset(self, other: 'BinaryTree') -> bool:
+        for val in self:
+            if val not in other:
+                return False
+        return True
 
-        methods_with_empty_root_handler = [
-            'pop', 'traverse', 'insert', 'delete', 'clear'
-        ]
+    def is_superset(self, other: 'BinaryTree') -> bool:
+        for val in other:
+            if val not in self:
+                return False
+        return True
 
-        if isinstance(attr, MethodType
-                      ) and attr_name not in methods_with_empty_root_handler:
+    def is_disjoint(self, other: 'BinaryTree') -> bool:
+        for val in other:
+            if val in self:
+                return False
+        return True
 
-            def root_checker(*args, **kwargs):
-                if self.root.value is None:
-                    return None
-                retval = attr(*args, **kwargs)
-                return retval
+    def difference(self, other: 'BinaryTree') -> 'BinaryTree':
+        return self - other
 
-            return root_checker
+    def difference_update(self, other: 'BinaryTree') -> None:
+        self -= other
 
-        return attr
+    def intersection(self, other: 'BinaryTree') -> 'BinaryTree':
+        common_val = [val for val in self if val in other]
+        return self.fill_tree(common_val)
 
-    def __setattr__(self, attr_name, val):
-        if attr_name == '_node_type':
-            raise ValueError(f'{attr_name} of the tree cannot be altered!')
-        super().__setattr__(attr_name, val)
+    def intersection_update(self, other: 'BinaryTree') -> None:
+        common_val = [val for val in self if val in other]
+        self.root = self.fill_tree(common_val).root
 
     def __getitem__(self, key):
         mod_key = len(self) - abs(key) if key < 0 else key
@@ -453,7 +405,7 @@ class BinaryTree(Generic[CT]):
         return self.root.find_node(value)
 
     def __bool__(self) -> bool:
-        return self.root.value
+        return self.root.value is not None
 
     def __str__(self):
         return str(self.traverse())
